Avoid rsync failure in standby controller
When both controllers are rebooted at same time,
sometimes active controller is not pingable from
the standby controller causing rsync failure (luksdata).

So, in this commit added following conditions to check
before executing the rsync. Also, rsync will be retried
for 10 secs.
Cond 1: rsync pull should happen only if active controller is
        reachable/pingable.
Cond 2: rsync pull should happen only when both the
        controllers are on same sw_version.

By adding this fix, the rsync won't execute if active controller
is not reachable. Thus, avoid abrupt stopping of execution of
puppet manifest.
In this scenario, consistency of the luksdata between
active and standby will be assured by the push mechanism
on the active controller. And this pull mechanism is
useful in scenarios where new standby nodes are getting deployed.

Test Plan:
PASSED: Deploy ISO on AIO-SX, AIO-DX.
PASSED: System deployment is successful.
PASSED: Rebooted standby controller, rsync
        is executed on standby.
PASSED: system host-swact, rsync is executed
        successfully from standby controller.

Story: 2010873
Task: 49389

+++ puppet-manifests/src/modules/platform/manifests/filesystem.pp
@@ -627,6 +627,7 @@ class platform::filesystem::luks {
     exec { 'rsync_luks_folder':
       command   => '/usr/bin/rsync -v -acv --delete rsync://controller/luksdata/ /var/luks/stx/luks_fs/controller/',
       logoutput => true,
+      onlyif    => [ "test ${::controller_sw_versions_match} = true", '/usr/local/bin/connectivity_test -t 10 controller', ],
     }
   }
 }
